chore:relocate ThemeProvider
